refactor some entities and solve some problems

+++ src/main/java/com/staketab/minanames/entity/PayableTransactionEntity.java
@@ -27,5 +27,5 @@ public class PayableTransactionEntity {
     private Set<DomainEntity> domains;
 
     @Column(name = "amount")
-    private Double txAmount;
+    private Long txAmount;
 }
